Combine DerivationManager and ParseTree into DerivationBranch
