+++ kataja/parser/deprecated/LayeredParser.py
@@ -60,7 +60,7 @@ class LayeredParser(BaseParser):
         if right:
             self.forest.connect_node(parent=node, child=f.get_node(right), direction=g.RIGHT)
         node.update_label()
-        f.derivation_steps.save_and_create_derivation_step()
+        f.get_derivation_branch().save_and_create_derivation_step()
         return constituent
 
     def _new_constituent(self, args):
