+++ kataja/saved/DerivationStep.py
@@ -22,10 +22,8 @@
 #
 # ############################################################################
 
-from kataja.singletons import ctrl, log
 from kataja.SavedObject import SavedObject
 from kataja.SavedField import SavedField
-from kataja.syntactic_state_to_nodes import syntactic_state_to_nodes
 from kataja.syntax.SyntaxState import SyntaxState
 
 
@@ -40,7 +38,7 @@ class DerivationStep(SavedObject):
     """ Saveable SyntaxState.
      """
 
-    def __init__(self, syn_state, uid=None):
+    def __init__(self, syn_state, uid=None, parent_uid=None):
         super().__init__(uid=uid)
         if syn_state:
             self.tree_roots = syn_state.tree_roots
@@ -60,6 +58,7 @@ class DerivationStep(SavedObject):
             self.semantic_hierarchies = []
             self.iteration = 0
             self.log = []
+        self.parent_uid = parent_uid
         self.frozen = None
 
     def __str__(self):
@@ -81,7 +80,7 @@ class DerivationStep(SavedObject):
                 else:
                     print('cannot save open reference object ', obj)
         assert (c < max_depth)  # please raise the max depth if this is reached
-        self.frozen = (self.uid, data, self.msg, self.iteration)
+        self.frozen = (self.uid, data, self.msg, self.iteration, self.parent_uid)
 
     def to_syn_state(self):
         return SyntaxState(tree_roots=self.tree_roots, numeration=self.numeration, msg=self.msg,
@@ -103,105 +102,3 @@ class DerivationStep(SavedObject):
     semantic_hierarchies = SavedField("semantic_hierarchies")
     iteration = SavedField("iteration")
     log = SavedField("log")
-
-
-class DerivationStepManager(SavedObject):
-    """ Stores derivation steps for one forest and takes care of related
-    logic """
-
-    def __init__(self, forest=None):
-        super().__init__()
-        self.forest = forest
-        self.activated = False
-        self.current = None
-        self.derivation_steps = []
-        self.derivation_step_index = None
-
-    def save_and_create_derivation_step(self, syn_state: SyntaxState):
-        """ Derivation steps store syntax states. Nodes etc. will be
-        then updated and created as necessary the fly.
-        :param syn_state: SyntaxState-instance
-        :return:
-        """
-        d_step = DerivationStep(syn_state)
-        # Use Kataja's save system to freeze objects into form where they can be stored and restored
-        # without further changes affecting them.
-        d_step.freeze()
-        self.derivation_steps.append(d_step.frozen)
-
-    def save_derivation_step(self, derivation_step: DerivationStep):
-        if not derivation_step.frozen:
-            derivation_step.freeze()
-        self.derivation_steps.append(derivation_step.frozen)
-
-    def remove_iterations(self, iterations):
-        """ There may be derivation steps that have been created during the parse but are
-        found redundant or not interesting and that they should be removed. They can be
-        removed by their iteration index
-        :param iterations: list of iterations, all of these will be removed if found.
-        :return:
-        """
-        print(f'removing {len(iterations)} iterations.')
-        self.derivation_steps = [(uid, data, msg, i) for uid, data, msg, i
-                                 in self.derivation_steps
-                                 if i not in iterations]
-
-    #    @time_me
-    def restore_derivation_step(self):
-        if self.derivation_steps:
-            uid, frozen_data, msg, i = self.derivation_steps[self.derivation_step_index]
-            d_step = DerivationStep(None, uid=uid)
-            d_step.load_objects(frozen_data)
-            self.activated = True
-            self.current = d_step
-            syntactic_state_to_nodes(self.forest, d_step.to_syn_state())
-            if d_step.msg:
-                log.info(f'<b>msg: {d_step.msg}</b>')
-            for log_msg in d_step.log:
-                if log_msg.strip():
-                    log_msg = log_msg.replace("\t", "&nbsp;&nbsp;")
-                    log.info(f'<font color="#859900">{log_msg}</font>')
-
-    def get_derivation_step(self, index):
-        if self.derivation_steps:
-            uid, frozen_data, msg, i = self.derivation_steps[index]
-            d_step = DerivationStep(None, uid=uid)
-            d_step.load_objects(frozen_data)
-            return d_step
-
-    def next_derivation_step(self):
-        """
-        :return:
-        """
-        if self.derivation_step_index is None or self.derivation_step_index + 1 >= len(self.derivation_steps):
-            self.derivation_step_index = 0
-        else:
-            self.derivation_step_index += 1
-        self.restore_derivation_step()
-
-    def previous_derivation_step(self):
-        """
-        :return:
-        """
-        if self.derivation_step_index is None or self.derivation_step_index == 0:
-            self.derivation_step_index = len(self.derivation_steps) - 1
-        else:
-            self.derivation_step_index -= 1
-        self.restore_derivation_step()
-
-    def jump_to_derivation_step(self, i):
-        """
-        :return:
-        """
-        self.derivation_step_index = i
-        self.restore_derivation_step()
-
-    # ############## #
-    #                #
-    #  Save support  #
-    #                #
-    # ############## #
-
-    derivation_steps = SavedField("derivation_steps")
-    derivation_step_index = SavedField("derivation_step_index")  # , watcher=ctrl.main.forest_changed)
-    forest = SavedField("forest")
